requirements.txt
asgiref==3.6.0
bleach==6.0.0
boto3==1.26.62
botocore==1.29.62
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
decorator==5.1.1
distlib==0.3.6
Django==4.1.4
django-cors-headers==3.13.0
django-quill==0.1.0
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-summernote==0.8.20.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dynaconf==3.1.11
filelock==3.8.0
gunicorn==20.1.0
hvac==1.0.2
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
loguru==0.6.0
MarkupSafe==2.1.2
Pillow==9.2.0
platformdirs==2.5.4
pycparser==2.21
pyhcl==0.4.4
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
validators==0.20.0
virtualenv==20.16.7
webencodings==0.5.1
whitenoise==6.1.0
win32-setctime==1.1.0
razorpay==1.3.0